complete role assign issue
