+++ src/main/java/com/hw/factory/BeanFactory.java
@@ -3,19 +3,15 @@ package com.hw.factory;
 import com.hw.service.ClientDetailsServiceImpl;
 import jdk.jfr.Description;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Primary;
-import org.springframework.core.Ordered;
 import org.springframework.core.env.Environment;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.oauth2.provider.ClientDetailsService;
-import org.springframework.security.oauth2.provider.OAuth2RequestFactory;
 import org.springframework.security.oauth2.provider.approval.ApprovalStore;
 import org.springframework.security.oauth2.provider.approval.TokenApprovalStore;
 import org.springframework.security.oauth2.provider.approval.TokenStoreUserApprovalHandler;
-import org.springframework.security.oauth2.provider.code.AuthorizationCodeServices;
 import org.springframework.security.oauth2.provider.code.InMemoryAuthorizationCodeServices;
 import org.springframework.security.oauth2.provider.request.DefaultOAuth2RequestFactory;
 import org.springframework.security.oauth2.provider.token.DefaultTokenServices;
@@ -24,9 +20,6 @@ import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenCo
 import org.springframework.security.oauth2.provider.token.store.JwtTokenStore;
 import org.springframework.security.oauth2.provider.token.store.KeyStoreKeyFactory;
 import org.springframework.stereotype.Component;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
 
 import javax.annotation.Resource;
 import java.security.KeyPair;
@@ -46,89 +39,77 @@ public class BeanFactory {
     }
 
     @Bean
-    FilterRegistrationBean corsConfiguration() {
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        CorsConfiguration configuration = new CorsConfiguration();
-        configuration.addAllowedOrigin("*");
-        configuration.addAllowedHeader("Authorization");
-        configuration.addAllowedHeader("Content-Type");
-        configuration.addAllowedHeader("Accept");
-        configuration.addAllowedHeader("Access-Control-Request-Method");
-        configuration.addAllowedHeader("x-requested-with");
-        configuration.addAllowedMethod("POST");
-        configuration.addAllowedMethod("PATCH");
-        configuration.addAllowedMethod("GET");
-        configuration.addAllowedMethod("DELETE");
-        configuration.addAllowedMethod("PUT");
-        configuration.addAllowedMethod("OPTIONS");
-        configuration.setMaxAge(3600L);
-        source.registerCorsConfiguration("/oauth/token", configuration);
-        source.registerCorsConfiguration("/oauth/token_key", configuration);
-        source.registerCorsConfiguration("/api/v1/client", configuration);
-        source.registerCorsConfiguration("/api/v1/clients", configuration);
-        source.registerCorsConfiguration("/api/v1/resourceOwner", configuration);
-        source.registerCorsConfiguration("/api/v1/resourceOwners", configuration);
-        source.registerCorsConfiguration("/api/v1/resourceOwner/**/pwd", configuration);
-        source.registerCorsConfiguration("/api/v1/authorize", configuration);
-        FilterRegistrationBean bean = new FilterRegistrationBean(new CorsFilter(source));
-        //NOTE:make sure oauth security check happen after cors filter
-        bean.setOrder(Ordered.HIGHEST_PRECEDENCE);
-        return bean;
-    }
-
-
-    @Bean
     public TokenStoreUserApprovalHandler userApprovalHandler(TokenStore tokenStore) {
+
         TokenStoreUserApprovalHandler handler = new TokenStoreUserApprovalHandler();
+
         handler.setTokenStore(tokenStore);
+
         handler.setRequestFactory(new DefaultOAuth2RequestFactory(clientDetailsService));
+
         handler.setClientDetailsService(clientDetailsService);
+
         return handler;
     }
 
     @Bean
     public ApprovalStore approvalStore(TokenStore tokenStore) throws Exception {
+
         TokenApprovalStore store = new TokenApprovalStore();
+
         store.setTokenStore(tokenStore);
+
         return store;
     }
 
 
     @Bean
     public TokenStore tokenStore(JwtAccessTokenConverter accessTokenConverter) {
+
         return new JwtTokenStore(accessTokenConverter);
     }
 
     @Bean
     public JwtAccessTokenConverter accessTokenConverter(KeyPair keyPair) {
+
         JwtAccessTokenConverter converter = new JwtAccessTokenConverter();
+
         converter.setKeyPair(keyPair);
+
         return converter;
     }
 
     @Bean
     public KeyPair getKeyPair() {
+
         KeyStoreKeyFactory keyStoreKeyFactory =
                 new KeyStoreKeyFactory(new ClassPathResource(env.getProperty("jwt.key-store")), env.getProperty("jwt.password").toCharArray());
+
         return keyStoreKeyFactory.getKeyPair(env.getProperty("jwt.alias"));
     }
 
     @Bean
     @Primary
     public DefaultTokenServices tokenServices(TokenStore tokenStore) {
+
         DefaultTokenServices defaultTokenServices = new DefaultTokenServices();
+
         defaultTokenServices.setTokenStore(tokenStore);
+
         defaultTokenServices.setSupportRefreshToken(true);
+
         return defaultTokenServices;
     }
 
     @Bean
-    public InMemoryAuthorizationCodeServices authorizationCodeServices(){
+    public InMemoryAuthorizationCodeServices authorizationCodeServices() {
+
         return new InMemoryAuthorizationCodeServices();
     }
 
     @Bean
-    private DefaultOAuth2RequestFactory defaultOAuth2RequestFactory(ClientDetailsServiceImpl clientDetailsService){
+    private DefaultOAuth2RequestFactory defaultOAuth2RequestFactory(ClientDetailsServiceImpl clientDetailsService) {
+
         return new DefaultOAuth2RequestFactory(clientDetailsService);
     }
 }
